move all logged out actions in their own file
